# backend/src/main/resources/openapi/felix-hotel-api.yaml
openapi: 3.0.3
info:
  title: Felix Hotel API
  description: >
    Contratto OpenAPI da cui vengono generati (a build-time, via
    openapi-generator-maven-plugin) sia i DTO di richiesta/risposta sia le
    interfacce API dei Controller: questo file e' la fonte di verita' per
    la FORMA dei DTO ("components/schemas") e per le rotte ("paths"). I
    Controller implementano le interfacce generate, valorizzando solo il
    corpo dei metodi (pattern Controller -> Service invariato).
  version: 0.1.0
paths:
...
  /api/tipologie-camera:

    get:
      tags:
        - TipologieCamera
      operationId: elencaTipologieCamera
      summary: Elenco paginato delle tipologie di camera
      description: >
        Restituisce le tipologie di camera offerte dalla struttura, in ordine
        alfabetico di nome e divise in pagine. **Pubblico**: e' il catalogo
        che il frontoffice mostra a chi visita il sito senza essersi
        registrato, quindi non richiede autenticazione.
        Risposta: busta ApiBaseResponsePaginated con un **array di
        TipologiaCameraResponse** in 'data' e i metadati di pagina in 'page'.
      parameters:
        - name: page
          in: query
          required: false
          description: Numero della pagina richiesta, 0-based
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: >
            Quanti elementi contiene una pagina. Il massimo non e' arbitrario
            ne' negoziabile: senza un tetto, una sola richiesta con un size
            enorme si farebbe restituire l'intera tabella in un colpo.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Pagina di tipologie ('data' = array di TipologiaCameraResponse)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiBaseResponsePaginated'
              example:
                status: 200
                message: Tipologie di camera recuperate
                timestamp: 2026-08-06T10:15:30Z
                data:
                  - id: 1
                    nome: Doppia Superior
                    descrizione: Camera doppia con vista sul giardino
                    capienzaMax: 2
                    prezzoNotte: 120.00
                page:
                  pageNumber: 0
                  pageSize: 20
                  totalElements: 1
                  totalPages: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - TipologieCamera
      operationId: creaTipologiaCamera
      summary: Crea una tipologia di camera
      description: >
        Aggiunge una tipologia al catalogo. Riservato agli **ADMIN**: e' il
        listino della struttura, non qualcosa che un cliente possa toccare.
        Il nome deve essere libero — il confronto ignora le maiuscole, quindi
        "Doppia" e "doppia" sono lo stesso nome e la seconda riceve 409.
        Risposta: busta ApiBaseResponse con **TipologiaCameraResponse** in 'data'.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TipologiaCameraRequest'
      responses:
        '201':
          description: Tipologia creata ('data' = TipologiaCameraResponse)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiBaseResponse'
              example:
                status: 201
                message: Tipologia di camera creata
                timestamp: 2026-08-06T10:15:30Z
                data:
                  id: 1
                  nome: Doppia Superior
                  descrizione: Camera doppia con vista sul giardino
                  capienzaMax: 2
                  prezzoNotte: 120.00
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  responses:
    BadRequest:
      description: >
        Richiesta non valida: violazione dei vincoli dichiarati sullo schema
        del body (@Valid), JSON malformato, tipo di parametro sbagliato, o
        una regola applicativa sull'input (es. consenso privacy non
        accettato). E' l'unico errore con 'data' valorizzato: quando
        l'errore viene dalla validazione dei campi, 'data' contiene una
        mappa nome-campo -> messaggio, cosi' il client puo' segnalare il
        singolo input sbagliato senza interpretare il messaggio testuale.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiBaseResponse'
          example:
            status: 400
            message: Dati non validi
            timestamp: 2026-08-05T10:15:30Z
            data:
              password: deve contenere almeno 8 caratteri
              email: deve essere un indirizzo email valido
    Unauthorized:
      description: >
        Autenticazione assente o non valida: token mancante, malformato,
        scaduto, account disattivato, oppure credenziali errate sul login.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiBaseResponse'
          example:
            status: 401
            message: Autenticazione richiesta
            timestamp: 2026-08-05T10:15:30Z
            data: null
    Forbidden:
      description: >
        Account autenticato ma senza i permessi necessari (ruolo
        insufficiente). Diverso dal 401: qui l'identita' e' nota, non basta.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiBaseResponse'
          example:
            status: 403
            message: Permessi insufficienti per questa operazione
            timestamp: 2026-08-05T10:15:30Z
            data: null
...
    Conflict:
      description: >
        La richiesta e' formalmente valida ma confligge con lo stato
        attuale della risorsa (es. email gia' registrata).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiBaseResponse'
          example:
            status: 409
            message: Email gia' registrata
            timestamp: 2026-08-05T10:15:30Z
            data: null
...
    InternalServerError:
      description: >
        Errore inatteso lato server. Il messaggio e' volutamente generico:
        il dettaglio dell'eccezione resta nei log e non viene esposto al
        client.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiBaseResponse'
          example:
            status: 500
            message: Si e' verificato un errore imprevisto
            timestamp: 2026-08-05T10:15:30Z
            data: null
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token JWT ottenuto da /api/auth/login, da inviare nell'header
        Authorization come "Bearer <token>".
  schemas:
...
    TipologiaCameraRequest:
      type: object
      description: >
        Dati di una tipologia di camera in scrittura (creazione e
        aggiornamento usano lo stesso schema: una PUT sostituisce l'intera
        risorsa, quindi i campi richiesti sono gli stessi della POST).
        I limiti dichiarati qui non sono decorativi: rispecchiano le colonne
        del database (vedi V1__init_schema.sql). Senza, un nome piu' lungo
        della colonna o un prezzo fuori scala arriverebbero fino a Postgres e
        tornerebbero al client come 500 invece che come 400.
      required:
        - nome
        - capienzaMax
        - prezzoNotte
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 100
          description: >
            Nome commerciale della tipologia. Deve essere unico a meno delle
            maiuscole (vincolo garantito da un indice unico su lower(nome)).
          example: Doppia Superior
        descrizione:
          type: string
          maxLength: 2000
          description: >
            Testo libero mostrato nella scheda del catalogo. La colonna e' un
            TEXT senza limite, ma il limite qui c'e' lo stesso: un campo di
            testo pubblico e senza tetto e' un modo per far scrivere megabyte
            a chiunque abbia il permesso di scrivere.
          example: Camera doppia con vista sul giardino, letto matrimoniale e balcone
        capienzaMax:
          type: integer
          minimum: 1
          description: Numero massimo di persone che possono alloggiare nella camera
          example: 2
        prezzoNotte:
          type: number
          minimum: 0
          maximum: 99999999.99
          description: >
            Prezzo per notte in euro. Provvisorio per disegno: e' un prezzo
            fisso, segnaposto finche' non esistera' un vero sistema di tariffe
            stagionali (vedi ANALISI_FUNZIONALE). Il massimo corrisponde a
            quanto entra nella colonna NUMERIC(10,2).
          example: 120.00
...
    ApiBaseResponse:
      type: object
      description: >
        Busta di risposta standard di TUTTI gli endpoint non paginati:
        status, messaggio, timestamp e il payload in 'data'. E' l'unica
        busta non paginata del progetto — non si dichiara un envelope per
        endpoint. Quale oggetto arrivi dentro 'data' e' indicato nella
        description del singolo endpoint (OpenAPI non ha i generics: non
        c'e' modo di scrivere qui "data: T" e tenere insieme un solo
        schema).
      required:
        - status
        - message
        - timestamp
      properties:
        status:
          type: integer
          description: Codice di stato HTTP
          example: 200
        message:
          type: string
          description: Messaggio leggibile, di successo o di errore
          example: Operazione completata con successo
        timestamp:
          type: string
          format: date-time
          description: Istante in cui e' stata generata la risposta
        data:
          type: object
          description: >
            Payload della risposta: l'oggetto (o l'array) richiesto, null
            in caso di errore. Il tipo concreto e' documentato nella
            description dell'endpoint che restituisce questa busta.
            Unica eccezione al "null in caso di errore": sugli errori di
            validazione (400) 'data' porta la mappa nome-campo ->
            messaggio dei campi rifiutati.
    ApiBaseResponsePaginated:
      description: >
        Come ApiBaseResponse, ma per gli endpoint il cui 'data' e' una
        pagina di risultati: aggiunge 'page' con i metadati di
        paginazione. Schema separato (invece di 'page' sempre presente ma
        nullable) cosi' le risposte non paginate non portano un campo
        inutile.
      allOf:
        - $ref: '#/components/schemas/ApiBaseResponse'
        - type: object
          properties:
            page:
              $ref: '#/components/schemas/PageInfo'
    PageInfo:
      type: object
      description: Metadati di paginazione
      properties:
        pageNumber:
          type: integer
          description: Numero di pagina corrente (0-based)
          example: 0
        pageSize:
          type: integer
          description: Dimensione della pagina
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Numero totale di elementi su tutte le pagine
          example: 134
        totalPages:
          type: integer
          description: Numero totale di pagine
          example: 7